Clean out Node subcontext
