--- Neos.EventSourcedContentRepository/Classes/Domain/Context/Node/NodeAggregatesNameIsAmbiguous.php
@@ -1,22 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Neos\EventSourcedContentRepository\Domain\Context\Node;
-
-/*
- * This file is part of the Neos.ContentRepository package.
- *
- * (c) Contributors of the Neos Project - www.neos.io
- *
- * This package is Open Source Software. For the full copyright and license
- * information, please view the LICENSE file which was distributed with this
- * source code.
- */
-use Neos\Flow\Exception;
-
-/**
- * The exception to be thrown if a node aggregate's name is ambiguous
- */
-class NodeAggregatesNameIsAmbiguous extends Exception
-{
-}
