+++ Neos.EventSourcedContentRepository/Classes/Domain/Context/NodeAggregate/CopyableAcrossContentStreamsInterface.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-namespace Neos\EventSourcedContentRepository\Domain\Context\Node;
+namespace Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate;
 
 /*
  * This file is part of the Neos.ContentRepository package.
