+++ Neos.EventSourcedNeosAdjustments/Classes/Ui/Domain/Model/Changes/Property.php
@@ -12,14 +12,12 @@ namespace Neos\EventSourcedNeosAdjustments\Ui\Domain\Model\Changes;
  * source code.
  */
 
-use Neos\ContentRepository\DimensionSpace\DimensionSpace\DimensionSpacePointSet;
 use Neos\ContentRepository\Domain\Model\NodeType;
 use Neos\ContentRepository\Domain\Projection\Content\TraversableNodeInterface;
 use Neos\ContentRepository\Domain\Service\NodeServiceInterface;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Command\DisableNodeAggregate;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Command\SetNodeProperties;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Command\EnableNodeAggregate;
-use Neos\EventSourcedContentRepository\Domain\Context\Node\NodeCommandHandler;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\NodeAggregateCommandHandler;
 use Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\NodeVariantSelectionStrategy;
 use Neos\EventSourcedContentRepository\Domain\ValueObject\PropertyValue;
@@ -84,12 +82,6 @@ class Property extends AbstractChange
      * @var bool
      */
     protected $isInline;
-
-    /**
-     * @Flow\Inject
-     * @var NodeCommandHandler
-     */
-    protected $nodeCommandHandler;
 
     /**
      * @Flow\Inject
@@ -200,7 +192,7 @@ class Property extends AbstractChange
      * @throws \Neos\ContentRepository\Exception\NodeException
      * @throws \Neos\ContentRepository\Exception\NodeTypeNotFoundException
      * @throws \Neos\EventSourcedContentRepository\Domain\Context\ContentStream\ContentStreamDoesNotExistYet
-     * @throws \Neos\EventSourcedContentRepository\Domain\Context\Node\NodeAggregatesTypeIsAmbiguous
+     * @throws \Neos\EventSourcedContentRepository\Domain\Context\NodeAggregate\Exception\NodeAggregatesTypeIsAmbiguous
      * @throws \Neos\EventSourcedContentRepository\Exception\DimensionSpacePointNotFound
      */
     public function apply()
@@ -227,7 +219,7 @@ class Property extends AbstractChange
                         ]
                     )
                 );
-                $this->nodeCommandHandler->handleSetNodeProperties($command)->blockUntilProjectionsAreUpToDate();
+                $this->nodeAggregateCommandHandler->handleSetNodeProperties($command)->blockUntilProjectionsAreUpToDate();
             } else {
                 // property starts with "_"
                 if ($propertyName === '_nodeType') {
@@ -248,10 +240,10 @@ class Property extends AbstractChange
                         $command = new EnableNodeAggregate(
                             $node->getContentStreamIdentifier(),
                             $node->getNodeAggregateIdentifier(),
-                            // TODO: what do we want to unhide? I.e. including NESTED dimensions?
-                            new DimensionSpacePointSet([$node->getOriginDimensionSpacePoint()])
+                            $node->getOriginDimensionSpacePoint(),
+                            NodeVariantSelectionStrategy::allSpecializations()
                         );
-                        $this->nodeCommandHandler->handleShowNode($command)->blockUntilProjectionsAreUpToDate();
+                        $this->nodeAggregateCommandHandler->handleEnableNodeAggregate($command)->blockUntilProjectionsAreUpToDate();
                     }
                 } else {
                     throw new \Exception("TODO FIX");
